requirements.txt
requests~=2.32.4
beautifulsoup4~=4.13.4
streamlit~=1.45.1
openai~=1.86.0
python-dotenv~=1.1.0